% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{calculate_alpha}
\alias{calculate_alpha}
\title{Get the alpha level}
\usage{
calculate_alpha(level, fill)
}
\arguments{
\item{level}{An integer level of the sunburst plot.}

\item{fill}{A character vector (or a factor) with the fill levels.}
}
\value{
A number between 0.5 and 1. The maximum value is 1 (lowest/innermost level) and the minimum value is 0.5 (highest/outermost level). 0 if the fill is \code{NA}.
}
\description{
A wrapper for \code{\link[=calculate_alpha_wo_fill]{calculate_alpha_wo_fill()}} with 0 alpha values for non filled parts.
}
\details{
The value of the fill is not important. The only importance of the argument is to check if the fill is missing or not.
}
