% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_alpha.R
\name{calculate_alpha_wo_fill}
\alias{calculate_alpha_wo_fill}
\title{Get the alpha level}
\usage{
calculate_alpha_wo_fill(level)
}
\arguments{
\item{level}{An integer level of the sunburst plot.}
}
\value{
A number between 0.5 and 1. The maximum value is 1 (lowest/innermost level) and the minimum value is 0.5 (highest/outermost level).
}
\description{
Maps the level of the plot to an alpha value.
}
