% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle.R
\name{calculate_angle}
\alias{calculate_angle}
\title{Calculate the label angle}
\usage{
calculate_angle(box_max, box_min)
}
\arguments{
\item{box_max}{A numeric vector of upper endpoints of segments as a proportion.}

\item{box_min}{A numeric vector of lower endpoints of segments as a proportion.}
}
\value{
The angle correction as should be given to ggplot.
}
\description{
Get the angle from the end points of the segment. The angles makes the segment labels perpendicular to the circle.
An angle of 0 is the positive x-axis so the angle adjustment is directly the correction for ggplot.
A correction is applied to make the labels on the left side of the plot more readable.
}
