% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ctcae_colors}
\alias{ctcae_colors}
\title{CTCAE Color Palette}
\usage{
ctcae_colors(include_grade_5 = TRUE)
}
\arguments{
\item{include_grade_5}{Include Grade 5?}
}
\value{
Named character vector of colors.
}
\description{
Returns the color palette for CTCAE grades.
}
