% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ctcae_labels}
\alias{ctcae_labels}
\title{CTCAE Grade Labels}
\usage{
ctcae_labels(include_grade_5 = TRUE)
}
\arguments{
\item{include_grade_5}{Include Grade 5?}
}
\value{
Named character vector of labels.
}
\description{
Returns standard labels for CTCAE grades.
}
