% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunplot_dataset.R
\name{get_sunplot_dataset}
\alias{get_sunplot_dataset}
\title{Get a prepared dataset to draw the sunburst plot.}
\usage{
get_sunplot_dataset(.dat, ..., .count_var)
}
\arguments{
\item{.dat}{The data frame or a tibble with the variables and the values for the plot. See the example for the input dataset structure.}

\item{...}{Variables in \code{.dat} that matches the order of the circles in the sunburst plot (innermost to outermost).}

\item{.count_var}{The variable with the values to decide the sizes of the circles.}
}
\value{
A tibble to provide as input to the \code{draw_sunburst_plot} function. See details for the list of variables.
}
\description{
Creates a dataset prepared as input to the \code{draw_sunburst_plot} function.
}
\details{
The returned dataset includes the following columns.
\itemize{
\item \code{.name}: Label of the segment.
\item \code{.level}: Level of the sunburst plot. Level 1 is the innermost level.
\item \code{.fill}: Value for the fill variable. This is not the final colour in the plot, but the same value will have the same colour for all levels.
\item \code{.box_min} & \code{.box_max}: Minimum and maximum bounds for the segment.
\item \code{.alpha}: The transparency for that level.
\item \code{.angle}: The angle of the text label.
}
}
\examples{
in_dat <- tibble::tibble(
level1_var = c(
rep("l1 A", times = 3),
rep("l1 B", times = 4)
),
level2_var = c(
"l2 A", "l2 A", "l2 B", 
"l2 C", "l2 C", "l2 D", "l2 D"
),
level3_var = c(
"l3 A", "l3 B", "l3 C", 
"l3 D", "l3 E", "l3 F", "l3 G"
),
values_var = c(
123, 456, 798, 
987, 654, 321, 159
)
)

in_dat


get_sunplot_dataset(
.dat = in_dat,
level1_var, level2_var, level3_var,
.count_var = values_var
)
}
