% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsunburst-package.R
\docType{package}
\name{ggsunburst-package}
\alias{ggsunburst-package}
\alias{ggsunburst}
\title{ggsunburst: Sunburst Charts for Hierarchical and Longitudinal Data}
\description{
Create sunburst charts using the ggplot2 framework.
Sunburst charts display hierarchical data as concentric rings, where each
ring represents a level in the hierarchy. This package is particularly
suited for visualizing changes in adverse event grades over time at the
individual patient level, but can be used for any hierarchical or
longitudinal categorical data.
}
\section{Main Functions}{

\itemize{
\item \code{\link{sunburst}}: Create a sunburst chart from hierarchical data
\item \code{\link{sunburst_ae}}: Specialized function for adverse event visualization
\item \code{\link{geom_sunburst}}: ggplot2 geom for sunburst segments
\item \code{\link{prepare_sunburst_data}}: Convert data to sunburst format
}
}

\section{Color Scales}{

\itemize{
\item \code{\link{scale_fill_ctcae}}: CTCAE grade color scale
\item \code{\link{scale_fill_severity}}: General severity color scale
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kestrel99/ggsunburst}
  \item Report bugs at \url{https://github.com/kestrel99/ggsunburst/issues}
}

}
\author{
\strong{Maintainer}: Justin Wilkins \email{justin.wilkins@occams.com}

}
