% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{highlight_nodes}
\alias{highlight_nodes}
\title{Highlight nodes in the tree}
\usage{
highlight_nodes(plot, data, fill = "grey82", color = "white", ymax = 0,
  node_names, alpha = 1)
}
\arguments{
\item{plot}{object created by any of the ggsunburst function \code{\link{ggtree}}}

\item{data}{object obtained using the \code{\link{sunburst_data}} function}

\item{fill}{fill color for highlight rectangle}

\item{color}{outline color for highlight rectangle}

\item{ymax}{rightmost coordinate for rectangle, default 0}

\item{node_names}{list of nodes to be highlighted}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Highlight nodes in the tree
}
