% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{icicle}
\alias{icicle}
\title{Creates icicle diagram using ggplot2.}
\usage{
icicle(data, rects.fill = "white", rects.fill.aes = 0,
  rects.color = "black", rects.size = 0.5, blank = T, leaf_labels = T,
  leaf_labels.size = 2, leaf_labels.color = "black", node_labels = F,
  node_labels.size = 2, node_labels.color = "black",
  node_labels.min = 0.25)
}
\arguments{
\item{data}{object obtained using the \code{\link{sunburst_data}} function}

\item{rects.fill}{color of space-filled nodes}

\item{rects.fill.aes}{color of space-filled nodes mapped to a variable}

\item{rects.color}{color of line delimiter between partitions}

\item{rects.size}{size of line delimiter between partitions}

\item{blank}{if TRUE, a blank theme is applied}

\item{leaf_labels}{if TRUE, shows leaf labels}

\item{leaf_labels.size}{size for text of labels}

\item{leaf_labels.color}{color for text of labels}

\item{node_labels}{if TRUE, shows leaf labels}

\item{node_labels.size}{size for text of labels}

\item{node_labels.color}{color for text of labels}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Creates icicle diagram using ggplot2.
}
\seealso{
\code{\link{sunburst_data}}
}
