% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_utils.R
\name{label_orientations}
\alias{label_orientations}
\title{Label Orientation Options}
\usage{
label_orientations()
}
\value{
A character vector of orientation options with descriptions.
}
\description{
Helper to describe available label orientation options.
}
\examples{
label_orientations()
}
