% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_utils.R
\name{make_sequence}
\alias{make_sequence}
\title{Create a sequence of the stages across states}
\usage{
make_sequence(df, patid = patid, state = state, stage = stage)
}
\arguments{
\item{df}{Input data}

\item{patid}{Subject Identifier}

\item{state}{State}

\item{stage}{Stage}
}
\value{
Data frame
}
\description{
This function creates a sequence
}
