% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_data.R
\name{nw_print}
\alias{nw_print}
\title{Prints the tree structure encoded in the newick input}
\usage{
nw_print(newick, format = 8, ladderize = F, ultrametric = F)
}
\arguments{
\item{newick}{either a string or a file, in newick format (see https://pythonhosted.org/ete2/tutorial/tutorial_trees.html#reading-and-writing-newick-trees)}

\item{ladderize}{if not FALSE, sort the partitions of a given tree according to their size. 'left' to invert order}

\item{ultrametric}{converts a tree to ultrametric topology (all leaves must have the same distance to root)}
}
\description{
Prints the tree structure encoded in the newick input
}
