% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trajectory.R
\name{prepare_ae_trajectory}
\alias{prepare_ae_trajectory}
\title{Prepare Adverse Event Trajectory Data}
\usage{
prepare_ae_trajectory(data, subject, time, grade, time_order = NULL, ...)
}
\arguments{
\item{data}{A data frame with AE data.}

\item{subject}{Column name for subject identifier.}

\item{time}{Column name for timepoint/visit.}

\item{grade}{Column name for AE grade.}

\item{time_order}{Order of timepoints (inner to outer).}

\item{...}{Additional arguments passed to \code{prepare_trajectory_data}.}
}
\value{
A data frame ready for trajectory sunburst plotting.
}
\description{
Convenience function for preparing AE grade trajectory data.
}
