% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_data.R
\name{sunburst_data}
\alias{sunburst_data}
\title{Extracts data from the newick format in a list of data frames.}
\usage{
sunburst_data(input, type = "newick", sep = NULL, ladderize = F,
  ultrametric = F, xlim = 360, rot = 0, node_attributes = "")
}
\arguments{
\item{ladderize}{if not FALSE, sort the partitions of a given tree according to their size. 'left' to invert order}

\item{ultrametric}{converts a tree to ultrametric topology (all leaves must have the same distance to root)}

\item{newick}{either a string or a file, in newick format (see https://pythonhosted.org/ete2/tutorial/tutorial_trees.html#reading-and-writing-newick-trees)}
}
\value{
a list of data frames
}
\description{
This functions extracts relevant information from
a tree structure to be plotted as an adjacency diagram (sunburst and icicle)
}
\seealso{
\code{\link{sunburst}}

\code{\link{icicle}}
}
