% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tile}
\alias{tile}
\title{Adds a geom_tile layer.}
\usage{
tile(plot, data, variables, color = "white", height = 1, width = 1,
  add_to_y = 0, xlabels = 1, xfactor = 1.1, angle = 90, text.size = 3,
  labels = F)
}
\arguments{
\item{plot}{object created by any of the ggsunburst functions: \code{\link{sunburst}}, \code{\link{ggtree}} or \code{\link{icicle}}.}

\item{data}{data frame. It must contain \code{x} and \code{y} variables, plus the other variables passed as parameter \code{variables}.}

\item{variables}{vector of variables to be included in the tile.}

\item{color}{color passed to \code{geom_tile}.}

\item{height}{height of the tiles.}

\item{width}{width of the tiles.}

\item{add_to_y}{move the tile along the \code{y} axis.}

\item{xlabels}{\code{x} mapping for labels.}

\item{xfactor}{xfactor.}

\item{angle}{angle for labels text.}

\item{text.size}{size for labels text.}

\item{labels}{if set to \code{TRUE}, variables names are shown.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Adds a geom_tile layer.
}

