% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{ggtree}
\alias{ggtree}
\title{Creates tree diagram using ggplot2.}
\usage{
ggtree(data = 0, plot = 0, rotate = T, color = "black", size = 1,
  blank = T, labels = T, text.size = 3, text.color = "black",
  show_scale = F, show_scale_length = 0, polar = F)
}
\arguments{
\item{data}{object obtained using the \code{\link{sunburst_data}} function}

\item{plot}{a \code{\link[ggplot2]{ggplot}} object}

\item{rotate}{if TRUE (default), the root appears on the left}

\item{color}{color of the segments (branches)}

\item{size}{size of the segments (branches)}

\item{blank}{if TRUE, a blank theme is applied}

\item{labels}{if TRUE, leaf labels are shown}

\item{text.size}{size for text of labels}

\item{text.color}{color for text of labels}

\item{show_scale}{show branch length scale}

\item{polar}{show circular tree}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Creates tree diagram using ggplot2.
}
\seealso{
\code{\link{sunburst_data}}
}
